pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prism-review"
version = "0.1.0"
description = "AI-powered PR code review CLI"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "openai>=1.75",
    "PyGithub>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
prism = "prism_review.cli:main"
